deployment/drift_detection/requirements.txt
evidently
pandas
psycopg2
requests
Flask==2.3.3
google-cloud-storage